Add configurable GOR modes and smoke training example

+++ sentence_transformers/sentence_transformer/evaluation/embedding_similarity.py
@@ -169,8 +169,8 @@ class EmbeddingSimilarityEvaluator(BaseEvaluator):
         embeddings2 = self.embed_inputs(model, self.sentences2)
         # Binary and ubinary embeddings are packed, so we need to unpack them for the distance metrics
         if self.precision == "binary":
-            embeddings1 = (embeddings1 + 128).astype(np.uint8)
-            embeddings2 = (embeddings2 + 128).astype(np.uint8)
+            embeddings1 = (embeddings1.astype(np.int16) + 128).astype(np.uint8)
+            embeddings2 = (embeddings2.astype(np.int16) + 128).astype(np.uint8)
         if self.precision in ("ubinary", "binary"):
             embeddings1 = np.unpackbits(embeddings1, axis=1)
             embeddings2 = np.unpackbits(embeddings2, axis=1)
